Rename credential classes to match what they do.
I had named the files and classes incorrectly.
