+++ storage/client/storage_client_unit_tests.bzl
@@ -2,10 +2,11 @@
 storage_client_unit_tests = [
     "bucket_metadata_test.cc",
     "bucket_test.cc",
+    "internal/authorized_user_credentials_test.cc",
     "internal/default_client_test.cc",
+    "internal/google_application_default_credentials_file_test.cc",
     "internal/nljson_test.cc",
     "internal/parse_rfc3339_test.cc",
-    "internal/service_account_credentials_test.cc",
     "link_test.cc",
 ]
